requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pillow==9.5.0
numpy==1.25.2
requests==2.31.0 